Separation of GUI and Code
However, randomization of board does not work properly.
What it is doing- every time a person registers, a new Game Board is
created, each randomized.
How it should work - each Game Board should be the same dimensions. It
is acceptable to create an instance of a Game Board after each player
registers. There is a preliminary window that requires the number of
players to be between 2 and 4.  Without doing so the players cannot
progress to the registration menu.  I like this functionality. In case a
player puts down that 3 people want to play the game but then a 4th
person comes along, they will still be able to register the player
without having to exit the program and start over (as there is no back
option).